+++ src/pages/Home/components/ClothesSection/loading.tsx
@@ -4,9 +4,9 @@ import styled from '@emotion/styled';
 const RecommendClothesLoading = () => {
   return (
     <RecommendClothesWrap>
-      <Skeleton variant='rounded' height={96} />
-      <Skeleton variant='rounded' height={96} />
-      <Skeleton variant='rounded' height={96} />
+      <Skeleton variant='rounded' height={93} />
+      <Skeleton variant='rounded' height={93} />
+      <Skeleton variant='rounded' height={93} />
     </RecommendClothesWrap>
   );
 };
@@ -17,6 +17,5 @@ const RecommendClothesWrap = styled.div`
   display: flex;
   flex-direction: column;
   gap: 16px;
-  padding-bottom: 16px;
   background-color: ${({ theme }) => theme.colors.blueGrey[100]};
 `;
